+++ index.ts
@@ -10,21 +10,6 @@ import {stringify, stringifyError, parse, parseError} from './lib/serialization'
 
 const DEFAULT_RPC_EXPIRY = 15000; // 15 seconds
 
-function maybeFromCallback<T>(
-  fn:
-    ((done: Rabbitr.Callback<T>) => void) |
-    (() => PromiseLike<T>)
-): Bluebird<T> {
-  let callback: Rabbitr.Callback<T>;
-  let promise = Bluebird.fromCallback<T>(_callback => (callback = _callback) && void 0);
-
-  let val = (fn as Function)(callback);
-  if (val && val.then) {
-    return Bluebird.resolve(val);
-  }
-  return promise;
-}
-
 function bluebirdUnwrapRejection<T>(rejection: T): T;
 function bluebirdUnwrapRejection(rejection: any): any {
   if (rejection && rejection.isOperational && rejection.cause) {
@@ -85,7 +70,7 @@ class Rabbitr {
     if (!this.eventListeners[eventName]) {
       throw new Error(`Triggering an event without a listener: ${eventName}`);
     }
-    return maybeFromCallback<void>(this.eventListeners[eventName].bind(null, message));
+    return this.eventListeners[eventName](message);
   }
 
   private removeAllListeners(): void {
@@ -212,7 +197,7 @@ class Rabbitr {
           this.connected = true;
 
           this.log('ready');
-          return maybeFromCallback<void>(this.opts.setup || (() => Bluebird.resolve()))
+          return Bluebird.resolve(this.opts.setup ? this.opts.setup() : null)
             .then(() => {
               this.ready = true;
               return conn;
@@ -226,8 +211,8 @@ class Rabbitr {
   }
 
   // istanbul ignore next
-  public whenReady(callback?: () => void): Bluebird<void> {
-    return this.connectionPromise.then<void>(() => void 0).asCallback(callback);
+  public whenReady(): Bluebird<void> {
+    return this.connectionPromise.then<void>(() => void 0);
   }
 
   private _formatName(name: string): string {
@@ -241,7 +226,7 @@ class Rabbitr {
 
   private destroyPromise: Bluebird<void>;
   /** method to destroy anything for this instance of rabbitr */
-  destroy(cb?: Rabbitr.ErrorCallback): Bluebird<void> {
+  destroy(): Bluebird<void> {
     if (!this.destroyPromise) {
       this.log(`${red('destroying')}`);
       this.destroyPromise = Bluebird.each(this._openChannels, channel =>
@@ -271,34 +256,20 @@ class Rabbitr {
         );
       });
     }
-    return this.destroyPromise.asCallback(cb);
+    return this.destroyPromise;
   }
 
   // standard pub/sub stuff
 
-  // @deprecated signature
-  send(topic: string, data: any, cb?: Rabbitr.ErrorCallback, opts?: Rabbitr.ISendOptions): Bluebird<void>;
   send(topic: string, data: any, opts?: Rabbitr.ISendOptions): Bluebird<void>;
-  // @deprecated signature
-  send<TInput>(topic: string, data: TInput, cb?: Rabbitr.ErrorCallback, opts?: Rabbitr.ISendOptions): Bluebird<void>;
   send<TInput>(topic: string, data: TInput, opts?: Rabbitr.ISendOptions): Bluebird<void>;
 
-  send<TInput>(topic: string, data: TInput, cb?: Rabbitr.ErrorCallback, opts?: Rabbitr.ISendOptions): Bluebird<void> {
-    // istanbul ignore next
-    if (typeof opts === 'function') {
-      let tmp = cb;
-      cb = opts as any;
-      opts = tmp as any;
-    } else if (cb && typeof cb !== 'function') {
-      opts = cb as any;
-      cb = null;
-    }
-
+  send<TInput>(topic: string, data: TInput, opts?: Rabbitr.ISendOptions): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.send(topic, data, cb, opts)
+        this.send(topic, data, opts)
       );
     }
 
@@ -316,25 +287,18 @@ class Rabbitr {
           headers: opts && opts.headers,
         }
       );
-    }).asCallback(cb);
+    });
   }
 
-  subscribe(topic: string, cb?: Rabbitr.Callback<any>): Bluebird<void>;
-  subscribe(topic: string, opts?: Rabbitr.ISubscribeOptions, cb?: Rabbitr.Callback<any>): Bluebird<void>;
-  subscribe<TMessage>(topic: string, cb?: Rabbitr.Callback<TMessage>): Bluebird<void>;
-  subscribe<TMessage>(topic: string, opts: Rabbitr.ISubscribeOptions, cb?: Rabbitr.Callback<TMessage>): Bluebird<void>;
+  subscribe(topic: string, opts?: Rabbitr.ISubscribeOptions): Bluebird<void>;
+  subscribe<TMessage>(topic: string, opts?: Rabbitr.ISubscribeOptions): Bluebird<void>;
 
-  subscribe<TMessage>(topic: string, opts?: Rabbitr.ISubscribeOptions, cb?: Rabbitr.ErrorCallback): Bluebird<void> {
-    if (typeof opts === 'function') {
-      cb = opts as any;
-      opts = null;
-    }
-
+  subscribe<TMessage>(topic: string, opts?: Rabbitr.ISubscribeOptions): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.subscribe(topic, opts, cb)
+        this.subscribe(topic, opts)
       );
     }
 
@@ -435,16 +399,16 @@ class Rabbitr {
         return Bluebird.fromCallback(callback =>
           channel.consume(this._formatName(topic), processMessage, {}, callback)
         );
-      }).asCallback(cb);
+      });
     });
   }
 
-  bindExchangeToQueue(exchange: string, queue: string, cb?: Rabbitr.ErrorCallback): Bluebird<void> {
+  bindExchangeToQueue(exchange: string, queue: string): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.bindExchangeToQueue(exchange, queue, cb)
+        this.bindExchangeToQueue(exchange, queue)
       );
     }
 
@@ -461,7 +425,7 @@ class Rabbitr {
       ).then(ok => {
         return Bluebird.fromCallback(callback => channel.close(callback));
       });
-    }).asCallback(cb);
+    });
   };
 
   // timed queue stuff
@@ -469,19 +433,12 @@ class Rabbitr {
     return `dlq.${topic}.${uniqueID}`;
   }
 
-  setTimer<TData>(topic: string, uniqueID: string, data: TData, ttl: number, cb?: Rabbitr.ErrorCallback): Bluebird<void>;
-  setTimer<TData>(topic: string, uniqueID: string, data: TData, ttl: number, opts?: Rabbitr.ISetTimerOptions, cb?: Rabbitr.ErrorCallback): Bluebird<void>;
-  setTimer<TData>(topic: string, uniqueID: string, data: TData, ttl: number, opts?: Rabbitr.ISetTimerOptions, cb?: Rabbitr.ErrorCallback): Bluebird<void> {
-    if (typeof opts === 'function') {
-      cb = opts as any;
-      opts = null;
-    }
-
+  setTimer<TData>(topic: string, uniqueID: string, data: TData, ttl: number, opts?: Rabbitr.ISetTimerOptions): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.setTimer(topic, uniqueID, data, ttl, cb)
+        this.setTimer(topic, uniqueID, data, ttl)
       );
     }
 
@@ -504,15 +461,15 @@ class Rabbitr {
         expiration: `${ttl}`,
         headers: opts && opts.headers,
       });
-    }).asCallback(cb);
+    });
   }
 
-  clearTimer(topic: string, uniqueID: string, cb?: Rabbitr.ErrorCallback): Bluebird<void> {
+  clearTimer(topic: string, uniqueID: string): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.clearTimer(topic, uniqueID, cb)
+        this.clearTimer(topic, uniqueID)
       );
     }
 
@@ -522,7 +479,7 @@ class Rabbitr {
 
     return Bluebird.fromCallback(callback =>
       this._timerChannel.deleteQueue(timerQueue, {}, callback)
-    ).asCallback(cb);
+    );
   }
 
   // rpc stuff
@@ -553,23 +510,15 @@ class Rabbitr {
     });
   }
 
-  rpcExec(topic: string, data: any, cb?: Rabbitr.Callback<any>): Bluebird<any>;
-  rpcExec(topic: string, data: any, opts: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<any>): Bluebird<any>;
-  rpcExec<TInput, TOutput>(topic: string, data: TInput, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput>;
-  rpcExec<TInput, TOutput>(topic: string, data: TInput, opts: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput>;
+  rpcExec(topic: string, data: any, opts?: Rabbitr.IRpcExecOptions): Bluebird<any>;
+  rpcExec<TInput, TOutput>(topic: string, data: TInput, opts?: Rabbitr.IRpcExecOptions): Bluebird<TOutput>;
 
-  rpcExec<TInput, TOutput>(topic: string, data: TInput, opts?: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput> {
-    if (typeof opts === 'function') {
-      // shift arguments
-      cb = opts as any;
-      opts = null;
-    }
-
+  rpcExec<TInput, TOutput>(topic: string, data: TInput, opts?: Rabbitr.IRpcExecOptions): Bluebird<TOutput> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
-        this.rpcExec<TInput, TOutput>(topic, data, opts, cb)
+        this.rpcExec<TInput, TOutput>(topic, data, opts)
       );
     }
 
@@ -661,29 +610,19 @@ class Rabbitr {
           throw error;
         });
       });
-    }).asCallback(cb);
+    });
   }
 
-  rpcListener(topic: string, executor: Rabbitr.IRpcListenerExecutor<any, any>, callback?: Rabbitr.ErrorCallback): Bluebird<void>;
-  rpcListener(topic: string, opts: Rabbitr.IRpcListenerOptions<any, any>, executor: Rabbitr.IRpcListenerExecutor<any, any>, callback?: Rabbitr.ErrorCallback): Bluebird<void>;
-  rpcListener<TInput, TOutput>(topic: string, executor: Rabbitr.IRpcListenerExecutor<TInput, TOutput>, callback?: Rabbitr.ErrorCallback): Bluebird<void>;
-  rpcListener<TInput, TOutput>(topic: string, opts: Rabbitr.IRpcListenerOptions<TInput, TOutput>, executor: Rabbitr.IRpcListenerExecutor<TInput, TOutput>, callback?: Rabbitr.ErrorCallback): Bluebird<void>;
+  rpcListener(topic: string, opts: Rabbitr.IRpcListenerOptions<any, any>, executor: Rabbitr.IRpcListenerExecutor<any, any>): Bluebird<void>;
+  rpcListener<TInput, TOutput>(topic: string, opts: Rabbitr.IRpcListenerOptions<TInput, TOutput>, executor: Rabbitr.IRpcListenerExecutor<TInput, TOutput>): Bluebird<void>;
 
-  rpcListener<TInput, TOutput>(topic: string, opts: Rabbitr.IRpcListenerOptions<TInput, TOutput>, executor?, callback?: Rabbitr.ErrorCallback): Bluebird<void> {
+  rpcListener<TInput, TOutput>(topic: string, opts: Rabbitr.IRpcListenerOptions<TInput, TOutput>, executor?): Bluebird<void> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
       return this.whenReady().then(() =>
         this.rpcListener(topic, opts, executor)
       );
-    }
-
-    // istanbul ignore next
-    if ('function' === typeof opts) {
-      // shift arguments
-      callback = executor as any;
-      executor = opts as any;
-      opts = {};
     }
 
     var rpcQueue = this._rpcQueueName(topic);
@@ -750,7 +689,7 @@ class Rabbitr {
         skipMiddleware: true,
         durable: false,
       })
-    ).asCallback(callback);
+    );
   }
 
   // #region middleware
@@ -760,7 +699,7 @@ class Rabbitr {
     this.middlewareFn.push(fn);
   }
 
-  private useMiddleware(message: Rabbitr.IMessage<any>, next: () => PromiseLike<any>): Bluebird<any> {
+  private useMiddleware<T>(message: Rabbitr.IMessage<any>, next: () => T | PromiseLike<T>): Bluebird<T> {
     return Bluebird.try(
       this.middlewareFn.reduce<typeof next>(function(next, middlewareFn) {
         return middlewareFn.bind(null, message, next) as typeof next;
@@ -780,7 +719,7 @@ declare module Rabbitr {
     queuePrefix?: string;
 
     /** called once the connection is ready but before anything is bound (allows for ORM setup etc) */
-    setup?: ((done: Rabbitr.ErrorCallback) => void) | (() => PromiseLike<void>);
+    setup?: () => void | PromiseLike<void>;
     connectionOpts?: {
       heartbeat?: boolean;
     };
@@ -789,18 +728,8 @@ declare module Rabbitr {
     defaultRPCExpiry?: number;
   }
 
-  export interface ErrorCallback {
-    (err: Error): void;
-  }
-
-  export interface Callback<T> {
-    (err: Error): void;
-    (err: Error, data: T): void;
-  }
-
   export type IEventListener<TData> =
-    ((message: IMessage<TData>) => PromiseLike<void>) |
-    ((message: IMessage<TData>, respond: ErrorCallback) => void);
+    ((message: IMessage<TData>) => void | PromiseLike<void>);
 
   export interface Headers { [header: string]: string; }
 
@@ -813,8 +742,7 @@ declare module Rabbitr {
     prefetch?: number;
   }
   export type IRpcListenerExecutor<TInput, TOutput> =
-    ((message: IMessage<TInput>) => PromiseLike<TOutput>) |
-    ((message: IMessage<TInput>, respond: Callback<TOutput>) => void);
+    ((message: IMessage<TInput>) => TOutput | PromiseLike<TOutput>);
 
   export interface ISubscribeOptions {
     prefetch?: number;
@@ -834,15 +762,12 @@ declare module Rabbitr {
     data: TData;
     headers: { [header: string]: string; };
 
-    send(topic: string, data: any, cb?: Rabbitr.ErrorCallback, opts?: Rabbitr.ISendOptions): Bluebird<void>;
     send(topic: string, data: any, opts?: Rabbitr.ISendOptions): Bluebird<void>;
-    send<TInput>(topic: string, data: TInput, cb?: Rabbitr.ErrorCallback, opts?: Rabbitr.ISendOptions): Bluebird<void>;
     send<TInput>(topic: string, data: TInput, opts?: Rabbitr.ISendOptions): Bluebird<void>;
 
-    rpcExec(topic: string, data: any, cb?: Rabbitr.Callback<any>): Bluebird<any>;
-    rpcExec(topic: string, data: any, opts: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<any>): Bluebird<any>;
-    rpcExec<TInput, TOutput>(topic: string, data: TInput, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput>;
-    rpcExec<TInput, TOutput>(topic: string, data: TInput, opts: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput>;
+    rpcExec(topic: string, data: any): Bluebird<any>;
+    rpcExec(topic: string, data: any, opts: Rabbitr.IRpcExecOptions): Bluebird<any>;
+    rpcExec<TInput, TOutput>(topic: string, data: TInput, opts?: Rabbitr.IRpcExecOptions): Bluebird<TOutput>;
 
     isRPC: boolean;
     /** only for rpc: message headers to be sent back with the response */