+++ test/pubsub.ts
@@ -84,11 +84,9 @@ describe('rabbitr#pubsub', function() {
           )
       );
 
-    rabbit.on(queueName, function(message, reply) {
+    rabbit.on(queueName, function(message) {
       Bluebird
         .try(() => {
-          reply();
-
           // here we'll assert that the data is the same- the fact we received it means the test has basically passed anyway
           expect(JSON.stringify(testData)).to.equal(JSON.stringify(message.data));
         })
@@ -114,11 +112,9 @@ describe('rabbitr#pubsub', function() {
           )
       );
 
-    rabbit.on(queueName, function(message, reply) {
+    rabbit.on(queueName, function(message) {
       Bluebird
         .try(() => {
-          reply();
-
           expect(message.data).to.be.an.instanceOf(Buffer);
           expect(message.data.toString()).to.equal(data);
         })