+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [BREAKING] If you subscribe to a topic twice an error will be thrown.
 - [BREAKING] `ack` and `reject` on `IMessage` were removed. Either return a promise or use callbacks in `Rabbitr#on` (now second argument).
 - [BREAKING] middleware has been re-done from scratch. Check readme on how to use it.
+- [BREAKING] removed support for callbacks and updated most function signatures. You can maintain backwards compatibility by using `Bluebird#asCallback`, which is what rabbitr used to do anyway.
 - Add support for sending an receiving headers
 
 ## 8.4.0